finished the basic functionality interacting with the db
added a type converter because Romm does not accept reference types

added Repos package which contains all the async methods we gonna use to interact with our db

 Why async methods ?

           1) room does not allow you to access the db on the main thread

           2)using async methods accessing the db will give us a better ui that does not block each time we access some record in db

+++ app/src/main/java/com/ensao/mytime/Activityfeature/DataAccess/ActivityHistoryDAO.java
@@ -9,7 +9,7 @@ import com.ensao.mytime.Activityfeature.Busniss.ActivityHistory;
 public interface ActivityHistoryDAO {
 
     @Insert
-    int Insert(ActivityHistory Historyrecord);
+    long Insert(ActivityHistory Historyrecord);
 
 
     //we will work on this later
